+++ swim/macros/(Un-)Stun.js
@@ -61,6 +61,7 @@ async function main() {
         chatData += `, is no longer Stunned and looses Vulnerable after the turn.`;
         token.actor.update({ "data.status.isDistracted": false });
         token.actor.update({ "data.status.isStunned": false });
+        token.actor.update({ "data.status.isVulnerable": false });
         if (token.data.effects.includes(`${proneIconPath}`)) {
           token.toggleEffect(`${proneIconPath}`)
         };