+++ swim/macros/Token Vision.js
@@ -222,5 +222,5 @@ function main() {
       }
     }
   }).render(true);
+  // v.2.0.4
 }
-// v.2.0.3